public-gateway/pyproject.toml
[tool.poetry]
name = "public-gateway"
version = "0.1.0"
description = ""
authors = ["Alice <alice96ferreira@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
alembic = "^1.16.4"
celery = "^5.5.3"
pybreaker = "^1.4.0"
redis = "^6.4.0"
eventlet = "^0.40.2"
pydantic = {extras = ["email"], version = "^2.11.7"}
psycopg2-binary = "^2.9.10"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
kafka-python = "^2.2.15"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
black = "^25.1.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
